Move data_files, source_files, occupation_standards under admin namespace
