+++ app/controllers/users/sessions_controller.rb
@@ -24,7 +24,7 @@ class Users::SessionsController < Devise::SessionsController
   protected
 
   def after_sign_in_path_for(resource)
-    source_files_path
+    admin_source_files_path
   end
 
   # If you have extra params to permit, append them to the sanitizer.
