+++ app/views/admin/occupation_standards/_occupation_standard.html.erb
@@ -1,0 +1,32 @@
+<tr class="border-b">
+  <td class="text-sm text-gray-900 px-6 py-4 whitespace-nowrap text-left">
+    <%= link_to occupation_standard.title, [:admin, occupation_standard], class: "underline" %>
+  </td>
+  <td class="text-sm text-gray-900 px-6 py-4 whitespace-nowrap text-left">
+    <%= occupation_standard.occupation %>
+  </td>
+  <td class="text-sm text-gray-900 px-6 py-4 whitespace-nowrap text-left">
+    <%= occupation_standard.registration_agency %>
+  </td>
+  <td class="text-sm text-gray-900 px-6 py-4 whitespace-nowrap text-left">
+    <%= occupation_standard.occupation.onet_code || occupation_standard.onet_code %>
+  </td>
+  <td class="text-sm text-gray-900 px-6 py-4 whitespace-nowrap text-left">
+    <% if occupation_standard.importing? %>
+      <span class="pill bg-orange-500">
+        <%= occupation_standard.status.titleize %>
+      </span>
+    <% elsif occupation_standard.in_review? %>
+      <span class="pill bg-yellow-600">
+        <%= occupation_standard.status.titleize %>
+      </span>
+    <% elsif occupation_standard.published? %>
+      <span class="pill bg-green-700">
+        <%= occupation_standard.status.titleize %>
+      </span>
+    <% end %>
+  </td>
+  <td class="text-sm text-gray-900 px-6 py-4 whitespace-nowrap text-left">
+    <%= link_to "Edit", edit_admin_occupation_standard_path(occupation_standard) %>
+  </td>
+</tr>
