+++ app/views/admin/occupation_standards/_subsections.html.erb
