+++ app/views/admin/occupation_standards/index.html.erb
