+++ app/views/admin/occupation_standards/show.html.erb
@@ -1,0 +1,52 @@
+<div class="flex flex-col">
+  <div class="overflow-x-auto sm:-mx-6 lg:-mx-8">
+    <h1 class="w-full max-w-prose mx-auto text-center font-semibold text-xl mb-4">Occupation Standard for <%= @occupation_standard.title %></h1>
+    <div class="btn-primary align-right w-fit mb-2 ml-auto">
+      <%= link_to "Edit", edit_admin_occupation_standard_path(@occupation_standard) %>
+    </div>
+    <div class="border-t border-gray-200">
+      <dl>
+        <div class="bg-gray-50 px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
+          <dt>Status:</dt>
+          <dd><%= @occupation_standard.status.titleize %></dd>
+        </div>
+
+        <div class="bg-gray-50 px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
+          <dt "text-sm font-medium text-gray-500">Occupation:</dt>
+          <dd "mt-1 text-sm text-gray-900 sm:col-span-2 sm:mt-0"><%= @occupation_standard.occupation_title %></dd>
+        </div>
+
+        <div class="bg-white px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
+          <dt>URL:</dt>
+          <dd><%= @occupation_standard.url %></dd>
+        </div>
+
+        <div class="bg-gray-50 px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
+          <dt>Registration Agency:</dt>
+          <dd><%= @occupation_standard.registration_agency %></dd>
+        </div>
+
+        <div class="bg-white px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
+          <dt>RAPIDS code:</dt>
+          <dd><%= @occupation_standard.rapids_code %></dd>
+        </div>
+
+        <div class="bg-gray-50 px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
+          <dt>ONET code:</dt>
+          <dd><%= @occupation_standard.onet_code %></dd>
+        </div>
+
+        <div class="bg-white px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
+          <dt>Created at:</dt>
+          <dd><%= @occupation_standard.created_at %></dd>
+        </div>
+
+        <div class="bg-gray-50 px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
+          <dt>Updated at:</dt>
+          <dd><%= @occupation_standard.updated_at %></dd>
+        </div>
+      </dl>
+    </div>
+    <%= render partial: "subsections" %>
+  </div>
+</div>
