+++ app/views/admin/source_files/index.html.erb
@@ -1,0 +1,11 @@
+<div id="file-imports" class="w-full flex flex-col mx-auto">
+  <h1 class="w-full max-w-prose mx-auto text-center font-semibold text-xl mb-4">Source files</h1>
+  <% @source_files.find_each do |source_file| %>
+    <div class="flex justify-between items-center bg-white shadow-md rounded mb-4 py-3 px-3">
+      <p class="font-medium"><%= link_to source_file.filename, source_file.url %></p>
+      <p class="text-sm rounded-full py-2 px-4 bg-amber-300"><%= source_file.status.titleize %></p>
+      <%= link_to "Convert", admin_source_file_path(source_file) %>
+      <%= link_to "Edit", edit_admin_source_file_path(source_file) %>
+    </div>
+  <% end %>
+</div>
