--- app/views/data_imports/_form.html.erb
@@ -1,32 +1,0 @@
-<div class="flex flex-col w-full mx-auto">
-  <div class="w-full max-w-prose mx-auto bg-slate-300 my-5 p-3">
-  The file imported on this screen should be based on the <%= link_to "import template", "https://www.notion.so/888b37991598495cb22d0dabc08ae3b2?v=f29055b156fa471ea9c30e9467238e66", class: "default-colors", target: :_blank %>. Please review instructions for filling that template on the <%= link_to "ApprenticeshipStandards Notion", "https://www.notion.so/Instructions-060de1705e7d471fa8bee7a7c535a2d6", class: "default-colors", target: :_blank %>.
-  </div>
-  <div class="w-full max-w-prose mx-auto">
-    <%= form_with model: [@source_file, @data_import], class: "bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4" do |f| %>
-      <% if @data_import.errors.any? %>
-        <div class="mb-2">
-          <% @data_import.errors.full_messages.each do |msg| %>
-            <p><%= msg %></p>
-          <% end %>
-        </div>
-      <% end %>
-      <div class="mb-4">
-        <%= f.label :description, class: "form-label" %>
-        <%= f.text_field :description, class: "form-field" %>
-      </div>
-      <div class="mb-6">
-        <%= f.label :file, class: "form-label" %>
-        <%= f.file_field :file, class: "form-field" %>
-      </div>
-      <div class="mb-6">
-        <%= label_tag :last_file, "This is the last import for #{@source_file.filename}. Change its status to Completed after the import completes.", class: "form-label" %>
-        <%= check_box_tag :last_file %>
-      </div>
-      <div>
-        <%= f.submit "Submit", class: "btn-primary" %>
-        <%= link_to "Cancel", :back, class: "btn-secondary" %>
-      </div>
-    <% end %>
-  </div>
-</div>
