+++ app/views/layouts/_header.html.erb
@@ -9,8 +9,8 @@
   </div>
   <div class="w-full block flex-grow lg:flex lg:items-center lg:w-auto">
     <div class="text-md lg:flex-grow">
-      <%= link_to "Source files", source_files_path, class: "block mt-4 lg:inline-block lg:mt-0 text-blue-100 hover:text-white mr-4" %>
-      <%= link_to "Occupation Standards", occupation_standards_path, class: "block mt-4 lg:inline-block lg:mt-0 text-blue-100 hover:text-white mr-4" %>
+      <%= link_to "Source files", admin_source_files_path, class: "block mt-4 lg:inline-block lg:mt-0 text-blue-100 hover:text-white mr-4" %>
+      <%= link_to "Occupation Standards", admin_occupation_standards_path, class: "block mt-4 lg:inline-block lg:mt-0 text-blue-100 hover:text-white mr-4" %>
       <%= link_to "Edit account", edit_user_registration_path, class: "block mt-4 lg:inline-block lg:mt-0 text-blue-100 hover:text-white mr-4" %>
       <%= link_to "Sidekiq", sidekiq_path, class: "block mt-4 lg:inline-block lg:mt-0 text-blue-100 hover:text-white mr-4", target: :_blank %>
     </div>
