--- app/views/source_files/edit.html.erb
@@ -1,16 +1,0 @@
-<h1>Edit <%= @source_file.filename %></h1>
-
-<%= form_with model: @source_file, class: "bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4" do |f| %>
-  <div class="mb-4">
-    <%= f.label :status, class: "form-label" %>
-    <%= f.select :status, options_for_select(SourceFile.statuses.map { |k, v| [k.titleize, k] }), class: "form-field" %>
-  </div>
-  <div class="mb-4">
-    <%= f.label :metadata, class: "form-label" %>
-    <%= f.text_area :metadata, readonly: true, disabled: true %>
-  </div>
-  <div>
-    <%= f.submit "Update", class: "btn-primary" %>
-    <%= link_to "Cancel", :back, class: "btn-secondary" %>
-  </div>
-<% end %>
